Major refactoring of the polar polygon class.
No longer relies on separate side objects, but rather a list of booleans
describing which sides exist. This should result in a performance boost,
and also cleans up the code a large amount.

In addition this allows another change, which is to sync all the default
rotation angles of the game objects to the rotation of the centre polygon.
This ensures that all game objects (except the player) are always
aligned with the centre polygon, despite any timing issues that might happen.

Default value of OutlineColour for polarpolygons is now black.

Polar polygon outlines are now only drawn for outside edges, and not inside
ones. Possibly needs fixing to handle cases on the edge of open polygons.